migrate part of speech tests
